+++ src/discrete/detail/ContactResistance.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace nfx::silicon::discrete::detail
+{
+    inline constexpr float kOpenResistance = 1e12f;
+} // namespace nfx::silicon::discrete::detail